fix: fix make Error function
